HAN-45: moved sql exc handler to own file
